Price fields are int in the database

+++ src/SFA.DAS.Commitments.Database/Tables/ApprenticeshipFlexibleEmployment.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [dbo].ApprenticeshipFlexibleEmployment
 (
 	[ApprenticeshipId] BIGINT NOT NULL,
-	[EmploymentPrice] DECIMAL NOT NULL,
+	[EmploymentPrice] INT NOT NULL,
 	[EmploymentEndDate] DATETIME2 NOT NULL,
     CONSTRAINT [PK_FlexibleEmployment] PRIMARY KEY ([ApprenticeshipId])
 )
